apt and yum installs need sudo
it doesn't make sense to have a set of instructions where part of it sudo's and the rest doesn't when all of it requires root.  the edit's reflect that by adding missing sudo's so that the copy paste of these instructions doesn't fail.  because nobody should be logged in as root..