Generate Mermaid graph with "graph" command
Add a new --mermaid flag to the graph command and generate a Mermaid
graph when specified. The main usecase is embedding the generated text
to the Markdown. The Markdown renderer that supports Mermaid diagram can
take care of rendering the graph. That's why we only generate the text
and don't generate images like --dot flag.

I intentionally reduce the configuration knob for specifying the style.
The only knob we provide is the one that specifies the direction. This
is because there are an infinite amount of style requirements and
containerlab is not a tool for designing diagrams. Direction is helpful
for making a "baseline" structure. For example, when we write a DC network
topology, I know some people write server on top, but some people write
server at the bottom. This kind of style preference is common for network
domains so I think it's worth exposing direction as a flag to adjust those.

Why do we need to support Mermaid while we have support for Dot format?
That's because there are some use cases that we can't achieve with Dot,
but can achieve with Mermaid. For example, embedding a topology diagram
inside GitHub Markdown. Now GitHub natively supports rendering The
mermaid diagram on Markdown doesn't support Dot.

Another usecase is importing generated diagrams to draw.io. Sometimes a
bare Dot or Mermaid renderer doesn't draw the diagram that we want. In
such a case, we'd like to tweak the diagram a bit on the rich diagram
editor like draw.io, but without writing the entire diagram from
scratch. For such a case, importing a diagram into draw.io and editing
it helps. However, they only support the Mermaid format. No Dot support
at this moment.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>