Add "auto" option to ipv4-subnet and ipv6-subnet
If we do not care about the assigned IPv4/IPv6 subnet for the container
network, we let the container runtime allocate one for us. This
guarantees to avoid a collision of the hardcoded range (172.20.20.0/24
and 2001:172:20:20::/64) with any existing container networks.

This just works for the podman runtime. For the docker runtime we
compute the IPv6 subnet manually because it can only pick addresses from
the configured (global) subnet.