Fix yum and add DNF instructions for non-scripted installs
Yum instructions had inconsistent use of sudo and pointed to an older
path for the repo file.

The install script supports using dnf on Fedora, so it makes sense to also give the non-automated install commands.
Fedora 41 now ships with DNF5 which uses a slightly different syntax so both DNF 4 (Fedora versions <= 40) and DNF 5 (versions >= 41) are included