The correct one is below
/// tab | Geo coordinates

| Type     | Name                     |
| -------- | ------------------------ |
| `string` | `graph-geoCoordinateLat` |
| `string` | `graph-geoCoordinateLng` |

When using the geo positioning layout algorithm, you should set geographical coordinates in your topology definition so that TopoViewer is aware of where the nodes should be positioned.

If not defined the nodes will be randomly placed on the map.

TopoViewer accepts coordinates in latitude and longitude. Ensure to convert if using other coordinate systems (such as DMS).

`graph-geoCoordinateLat` sets the latitude component of the coordinates.
`graph-geoCoordinateLng` sets the longitude component of the coordinates.