Use HostConfig.Mounts for anonymous volumes
Containerlab passes all strings from the node "binds" property to Docker via
the HostConfig.Binds field. This API supports anonymous volumes and the
container does correctly start up with an new anonymous volume attached. But
the volume itself is missing the "Name" attribute for some reason (docker
inspect). This makes it impossible the then access the volume from other
containers, like with `docker run --volumes-from <existing-container> ...`.
An anonymous volume created with the new Mounts spec is for all intents and
purposes equivalent to one created with Binds, except it has the "Name"
attribute set to its random ID.

This has worked in the past, so maybe there is a regression in Docker?
Anyway by fixing it here we can ensure it will work for all versions.