Add volumes stanza, use Docker Mounts API
Containerlab now separates host binds from Docker volumes, with the use
of a new "volumes" stanza. This is a backwards-incompatible change from
the previous behavior where "binds" supported volumes (anonymous or
named). The "binds" property is now used only for binding host paths.

Since Docker v28, the HostConfig.Binds API has a regression where it
still creates an anonymous volume and attaches it to a container path,
but the volume is missing the "Name" attribute for some reason (docker
inspect). This makes it impossible to then access the volume from other
containers, like with `docker run --volumes-from <existing-container> ...`.
An anonymous volume created with the new HostConfig.Mounts spec is
equivalent to one created with Binds, except it has the "Name" attribute
set to its random ID.