feat: add SR-OS component config generation to SR-SIM (#2827)
This is currently only for distributed types, why?

1. Most integrated nodes have preprovisioned cards/mdas
2. Some DO NOT, which makes it hard when a user wants to only override
MDA, they should also defined card type.. but hold on a sec, overriden
MDAs are defined under slot A as when only slot A is defined we hop into
a special case where the env vars apply to the root node.. but in the
config we need to provision card 1, and a card-type for the card 1 IOM.

So that is a bit messy for integrated right now.. I will check what
types do/don't have a default config and work from there in another PR.

Have to test fully/add some tests for this.

closes #2823

---------

Co-authored-by: hellt <dodin.roman@gmail.com>