Sort components and add base node hostfile/ssh entry for SR-SIM (#2834)
If I have a topology like below. It will never complete deploying,
because the IP address is getting handed to the slot 1 , which is not
CPM. Since the CPM doens't have an IP, the deploy keeps waiting for the
CPM to get an IP (which never happens).

Maybe it is opinionated but I believe we should not force a sequential
ordering in the topo file.

This PR is sorting the components so that slot A should always get the
IP/host config and all component nodes attach to it's namespace.

```yaml
    sr1:
        kind: nokia_srsim
        image: nokia_srsim:25.7.r1
        license: srsim.txt
        type: sr-2s
        components:
            - slot: 1
              type: xcm-2s
              xiom:
              - mda:
                - slot: 1
                  type: ms18-100gb-qsfp28
                - slot: 2
                  type: ms18-100gb-qsfp28
                slot: 1
                type: iom-s-3.0t
            - slot: 2
              type: xcm-2s
              xiom:
              - slot: 2
                type: iom-s-3.0t
            - slot: A
              type: cpm-2s
            - slot: B
              type: cpm-2s
```
@hellt I would appreciate the 0.70.2 patch rel on this as well.

---------

Co-authored-by: hellt <dodin.roman@gmail.com>